% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataGroup.R
\name{getDataGroup}
\alias{getDataGroup}
\title{Separate data into different groups}
\usage{
getDataGroup(x, group)
}
\arguments{
\item{x}{A vector with the distance between individuals data. The vector contains the lower
triangle of a distance matrix filled in the same order as the lower.tri() function.}

\item{group}{A factor with the group to which each individual of the distance matrix belong.
The order of individuals on this variable must be the same as the distance matrix.}
}
\value{
It returns a list with number of elements equal to the number of levels on group,
i.e, the number of classes. Each element contains a vector with the distances between individuals
belonging to the same group. The elements are stuck on the same order as they were on vector x.
}
\description{
This function separate the distance between individuals data contained in x accordingly with 
his class on the variable group.
}
