% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelcor.R
\name{mantelcor}
\alias{mantelcor}
\title{Mantel correlation}
\usage{
mantelcor(D1, D2, D3, ...)
}
\arguments{
\item{D1}{Vector with the distance between individuals. Each element measures the distances in gene expression
at exon level between two individuals, of a particular gene.}

\item{D2}{The same as D1, but with the distances corresponding to another gene.}
}
\value{
Returns the partial correlation coefficient from the correlations between each pair of vectors
on the arguments.
}
\description{
The function performs a mantel correlation test between two data vectors. A third vector is
added to be used as a control parameter. It's a correction that consider the experiment conditions
in which the first two vectors where extracted.
}
