% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelcor2Group.R
\name{mantelcor2Group}
\alias{mantelcor2Group}
\title{Mantel test on two groups}
\usage{
mantelcor2Group(g1, g2, batchmat, group)
}
\arguments{
\item{g1}{The distance data vector for a particular gene.}

\item{g2}{The distance data vector for another gene.}

\item{batchmat}{The distance data vector containing the corrections due to experiment conditions.
The length of g1, g2 and batchmat must be the same.}

\item{group}{A factor with the group to which belongs each individual in the same order as they are on the matrix
from which the three first arguments belong. The function is performed so that the number of groups be 2 and the levels
of group be "A" & "B" in this order.}
}
\value{
The function calculate the absolute value of the difference in mantel correlation value from groups "A" & "B".
}
\description{
The function uses getDataGroup function to perform mantel test on data from individuals that
belongs to the same group. In this function two groups are considered: "A" & "B". The group "A" results from
joining the groups "AA" & "AB" from the factors used in mantelcorGroup. The group "B" is the group "BB".
}
