% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantelcorGroup.R
\name{mantelcorGroup}
\alias{mantelcorGroup}
\title{Mantel test on different groups}
\usage{
mantelcorGroup(g1, g2, batchmat, group)
}
\arguments{
\item{g1}{The distance data vector for a particular gene.}

\item{g2}{The distance data vector for another gene.}

\item{batchmat}{The distance data vector containing the corrections due to experiment conditions.
The length of g1, g2 and batchmat must be the same.}

\item{group}{A factor with the group to which belongs each individual in the same order as they are on the matrix
from which the three first arguments belong. The function is performed so that the number of groups be 3 and the levels
of group be "AA", "AB" & "BB" in this order.}
}
\value{
The function calculate the differences in mantel correlation value from each pair of groups and returns the 
maximum value between the absolute value of these differences.
}
\description{
The function uses getDataGroup function to perform mantel test on data from individuals that
belongs to the same group.
}
