% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAcor.R
\name{PCAcor}
\alias{PCAcor}
\title{PCA first component correlation}
\usage{
PCAcor(exprg1, exprg2, cov, ...)
}
\arguments{
\item{exprg1}{Dataframe with exon expression on a sample of individuals.}

\item{exprg2}{The same as exprg1 but with the exon expression of another gene. Notice the number of rows in exprg1 & exprg2
can differ as different genes has different number of exons.}

\item{cov}{The dataframe with the covariables due to conditions on the experiment.}
}
\value{
Returns the partial correlation coefficient from the scores of the first principal components that result from
each dataframe, using argument cov as the set of controlling variables.
}
\description{
The function performs a partial correlation test between two dataframes. Each dataframe contains the exon expression (rows)
of a particular gene in a sample of individuals (columns). A third dataframe with the same number of columns (individuals) is
added to be used as a control parameter. Its rows refer to covariables which affect the conditions of the experiment. On each 
data.frame individuals must appear on the same order.
}
