% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAcor2Group.R
\name{PCAcor2Group}
\alias{PCAcor2Group}
\title{PCA first component correlation by 3 groups}
\usage{
PCAcor2Group(exprg1, exprg2, cov, group)
}
\arguments{
\item{exprg1}{Dataframe with exon expression on a sample of individuals.}

\item{exprg2}{The same as exprg1 but with the exon expression of another gene. Notice the number of rows in exprg1 & exprg2
can differ as different genes has different number of exons.}

\item{cov}{The dataframe with the covariables due to conditions on the experiment.}
}
\value{
Returns the absolute value of the difference in partial correlation coefficient coming from each group.
}
\description{
The function performs the partial correlation test done in function PCAcor for two datasets and using argument cov as set of controlling variables.
This calculations are done by pulling away the individuals in two groups: "A" & "B".
}
